@@ -4,12 +4,12 @@ plugins {
 
 android {
     namespace 'vn.edu.usth.wikipedia'
-    compileSdk 34
+    compileSdk 35
 
     defaultConfig {
         applicationId "vn.edu.usth.wikipedia"
-        minSdk 24
-        targetSdk 34
+        minSdk 25
+        targetSdk 35
         versionCode 1
         versionName "1.0"
 
@@ -35,6 +35,10 @@ android {
 
 dependencies {
 
+    implementation libs.core.ktx
+    implementation libs.recyclerview
+    implementation libs.gson
+    implementation libs.okhttp
     implementation libs.appcompat
     implementation libs.material
     implementation libs.constraintlayout
@@ -42,10 +46,4 @@ dependencies {
     testImplementation libs.junit
     androidTestImplementation libs.ext.junit
     androidTestImplementation libs.espresso.core
-    implementation "androidx.core:core-ktx:1.6.0"  // Phiên bản thư viện Core KTX mới nhất
-    implementation "androidx.appcompat:appcompat:1.4.0"  // Phiên bản AppCompat mới nhất
-    implementation 'androidx.recyclerview:recyclerview:1.2.1'
-    implementation 'com.google.code.gson:gson:2.10.1'
-    implementation 'com.squareup.okhttp3:okhttp:5.0.0-alpha.13'
-    implementation 'androidx.appcompat:appcompat:1.5.0'
 }
