@@ -1,6 +1,7 @@
 package vn.edu.usth.wikipedia;
 
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -10,14 +11,16 @@ import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
+import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.activity.EdgeToEdge;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
 
+import vn.edu.usth.wikipedia.fragments.AccountInfoFragment;
 import vn.edu.usth.wikipedia.fragments.DonateFragment;
 import vn.edu.usth.wikipedia.fragments.LanguageFragment;
 import vn.edu.usth.wikipedia.fragments.LoginFragment;
@@ -25,9 +28,11 @@ import vn.edu.usth.wikipedia.fragments.LoginFragment;
 public class SettingActivity extends AppCompatActivity {
 
     Button loginButtonText, donateButtonText, languageButtonText;
-    ImageButton loginButton, donateButton, languageButton, closeButton;
+    ImageButton loginButton, donateButton, languageButton, closeButton, userImage;
     RelativeLayout settingBar;
     ImageView usthImage;
+    TextView usernameText;
+    SharedPreferences prefs;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -35,6 +40,10 @@ public class SettingActivity extends AppCompatActivity {
         EdgeToEdge.enable(this);
         setContentView(R.layout.activity_setting);
 
+        // Initialize shared preferences
+        prefs = getSharedPreferences("user_prefs", MODE_PRIVATE);
+
+        // Initialize views
         loginButton = findViewById(R.id.login_button);
         donateButton = findViewById(R.id.donate_button);
         languageButton = findViewById(R.id.language_button);
@@ -44,59 +53,36 @@ public class SettingActivity extends AppCompatActivity {
         closeButton = findViewById(R.id.close_more_button);
         settingBar = findViewById(R.id.setting_bar);
         usthImage = findViewById(R.id.usth_logo);
+        userImage = findViewById(R.id.user_image);
+        usernameText = findViewById(R.id.username_text);
 
-        closeButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
+        // Set username and login/logout button based on login status
+        boolean isLoggedIn = prefs.getBoolean("isLoggedIn", false);
+        String username = prefs.getString("username", "Guest");
+
+        if (isLoggedIn) {
+            loginButtonText.setText("Logout");
+            usernameText.setText(username); // Display logged-in username
+            loginButtonText.setOnClickListener(view -> logoutUser());
+        } else {
+            loginButtonText.setText("Login");
+            loginButtonText.setOnClickListener(view -> openLoginFragment());
+        }
+
+        // Set up button actions
+        userImage.setOnClickListener(view -> openAccountInfoFragment());
+        donateButton.setOnClickListener(view -> openDonateFragment());
+        donateButtonText.setOnClickListener(view -> openDonateFragment());
+        languageButton.setOnClickListener(view -> openLanguageFragment());
+        languageButtonText.setOnClickListener(view -> openLanguageFragment());
+        closeButton.setOnClickListener(v -> {
+            Intent intent = new Intent(SettingActivity.this, MainActivity.class);
+            startActivity(intent);
         });
 
-        loginButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                openLoginFragment();
-            }
-        });
-
-        donateButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                openDonateFragment();
-            }
-        });
-
-        languageButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                openLanguageFragment();
-            }
-        });
-
-        loginButtonText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                openLoginFragment();
-            }
-        });
-
-        donateButtonText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                openDonateFragment();
-            }
-        });
-
-        languageButtonText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                openLanguageFragment();
-            }
-        });
     }
 
-    public void openLoginFragment() {
-
+    private void openAccountInfoFragment() {
         loginButton.setVisibility(View.GONE);
         donateButton.setVisibility(View.GONE);
         languageButton.setVisibility(View.GONE);
@@ -105,6 +91,26 @@ public class SettingActivity extends AppCompatActivity {
         languageButtonText.setVisibility(View.GONE);
         settingBar.setVisibility(View.GONE);
         usthImage.setVisibility(View.GONE);
+        userImage.setVisibility(View.GONE);
+        usernameText.setVisibility(View.GONE);
+        Fragment accountInfoFragment = new AccountInfoFragment();
+        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+        transaction.replace(R.id.fragment_container, accountInfoFragment);
+        transaction.addToBackStack(null);
+        transaction.commit();
+    }
+
+    private void openLoginFragment() {
+        loginButton.setVisibility(View.GONE);
+        donateButton.setVisibility(View.GONE);
+        languageButton.setVisibility(View.GONE);
+        loginButtonText.setVisibility(View.GONE);
+        donateButtonText.setVisibility(View.GONE);
+        languageButtonText.setVisibility(View.GONE);
+        settingBar.setVisibility(View.GONE);
+        usthImage.setVisibility(View.GONE);
+        userImage.setVisibility(View.GONE);
+        usernameText.setVisibility(View.GONE);
         Fragment loginFragment = new LoginFragment();
         FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
         transaction.replace(R.id.fragment_container, loginFragment);
@@ -112,8 +118,7 @@ public class SettingActivity extends AppCompatActivity {
         transaction.commit();
     }
 
-    public void openDonateFragment() {
-
+    private void openDonateFragment() {
         loginButton.setVisibility(View.GONE);
         donateButton.setVisibility(View.GONE);
         languageButton.setVisibility(View.GONE);
@@ -122,15 +127,16 @@ public class SettingActivity extends AppCompatActivity {
         languageButtonText.setVisibility(View.GONE);
         settingBar.setVisibility(View.GONE);
         usthImage.setVisibility(View.GONE);
-        Fragment donanteFragment = new DonateFragment();
+        userImage.setVisibility(View.GONE);
+        usernameText.setVisibility(View.GONE);
+        Fragment donateFragment = new DonateFragment();
         FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
-        transaction.replace(R.id.fragment_container, donanteFragment);
+        transaction.replace(R.id.fragment_container, donateFragment);
         transaction.addToBackStack(null);
         transaction.commit();
     }
 
-    public void openLanguageFragment() {
-
+    private void openLanguageFragment() {
         loginButton.setVisibility(View.GONE);
         donateButton.setVisibility(View.GONE);
         languageButton.setVisibility(View.GONE);
@@ -139,11 +145,27 @@ public class SettingActivity extends AppCompatActivity {
         languageButtonText.setVisibility(View.GONE);
         settingBar.setVisibility(View.GONE);
         usthImage.setVisibility(View.GONE);
+        userImage.setVisibility(View.GONE);
+        usernameText.setVisibility(View.GONE);
         Fragment languageFragment = new LanguageFragment();
         FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
         transaction.replace(R.id.fragment_container, languageFragment);
         transaction.addToBackStack(null);
         transaction.commit();
+    }
+
+    private void logoutUser() {
+        SharedPreferences.Editor editor = prefs.edit();
+        editor.putBoolean("isLoggedIn", false);
+        editor.putString("username", "Guest");
+        editor.apply();
+
+        // Update UI after logout
+        loginButtonText.setText("Login");
+        usernameText.setText("Guest");
+        loginButtonText.setOnClickListener(view -> openLoginFragment());
+
+        Toast.makeText(this, "Logged out successfully", Toast.LENGTH_SHORT).show();
     }
 
     @Override
